Finished the option to compile as either static or dll.

+++ Source/GEOGL/ImGui/ImGuiLayer.cpp
@@ -107,7 +107,7 @@ namespace GEOGL{
 
         ImGuiIO& io = ImGui::GetIO();
         Application& app = Application::get();
-        io.DisplaySize = ImVec2(app.getWindow().getWidth(), app.getWindow().getHeight());
+        io.DisplaySize = ImVec2(static_cast<float>(app.getWindow().getWidth()), static_cast<float>(app.getWindow().getHeight()));
 
         // rendering
         ImGui::Render();
